==> McCray_PDOs/4_cycle_annotation.Rmd <==
---
title: "4.McCray_2019 (PDOs): cell cycle assignation"
author: "Romuald Parmentier"
date: "2024-03-26"
output: html_document
---


```{r Load libraries and create output folder, message = F}

# Libraries
library(Seurat)
library(SingleCellExperiment)
library(ggplot2)
library(dplyr)

# Functions and palettes
source(paste0(github_dir, "/Dolgos_Custom_Functions.R"))
source(paste0(github_dir, "/Dolgos_Custom_Color_Palettes.R"))

# Create output path

out_path = create_exp_folder(
  github_dir = github_dir,
  samples_ID = "McCray_PDOs",
  exp = "4_cycle_annotation"
)

```

# Prepare the data

```{r Load the files}

# Load sce after dimension reduction
file_path = get_exp_file_path(
  github_dir = github_dir,
  samples_ID = "McCray_PDOs",
  prev_exp = "3_dimension_reduction", 
  pattern = "sce_dim_reduced.rds")

sce = readRDS(file = file_path) 

```

# Cell type assignation 

Not povided by the authors and not needed here. 

# Cell cycle assignation 

Done according standrad Seurat procedure

```{r Cell cycle phase annotation}

# Transform sce object in seurat object
seurat = as.Seurat(
  x = sce,
  counts = "counts",
  data = NULL
)

DefaultAssay(object = seurat) = "originalexp"
seurat <- NormalizeData(seurat, normalization.method = "LogNormalize", scale.factor = 10000)
seurat <- ScaleData(seurat)

# Store cell cycles-associated genes in variables
s.genes <- cc.genes.updated.2019$s.genes
g2m.genes <- cc.genes.updated.2019$g2m.genes


# Compute the cell cycle signature score
seurat <- CellCycleScoring(seurat, s.features = s.genes, g2m.features = g2m.genes, set.ident = TRUE)

# Add cell cycle phase to the sce object
sce$cell_cycle_phase = seurat$Phase

# Print cell cycles phases stats
table(sce$cell_cycle_phase)

```

```{r Extracting metadata table}

md = as_tibble(colData(sce))

# get embeddings from 2 first PC
coords <- as_tibble(reducedDim(sce, "UMAP_on_PCA.1-20"))[,1:2]
colnames(coords) = c("UMAP_1","UMAP_2")

# combine dataframes
md <- cbind(md, coords)

```

```{r Visualizing cell cycle phase separately}

i = 1
plot_list = list()

for (cycle in unique(md$cell_cycle_phase)) {
  
  md_other <- subset(md, cell_cycle_phase != cycle)
  md_SPhase <- subset(md, cell_cycle_phase == cycle)
  
  plot <- ggplot() +
    # Rasterise the first layer
    ggrastr::rasterise(
      geom_point(
        data = md_other,
        aes(x = UMAP_1, y = UMAP_2), 
        fill = "grey", alpha = 0.2, size = 0.9, 
        shape = 21  # Ensure shape supports fill and color
      ), dpi = 200, scale = 0.5
    ) +
    # Rasterise the second layer
    ggrastr::rasterise(
      geom_point(
        data = md_SPhase,
        aes(x = UMAP_1, y = UMAP_2, fill = cell_cycle_phase), 
        alpha = 0.6, size = 1.5, color = "black", shape = 21
      ), dpi = 200, scale = 0.6
    ) +
    scale_fill_manual(values = pal.cycle) +
    theme_classic() +
    guides(
      alpha = "none",
      size = "none"
    )
  
  
  plot_list[[i]] = plot
  
  i = i + 1
  
}

multiple_page_layout <- gridExtra::marrangeGrob(
  grobs = plot_list, 
  nrow = 2, ncol = 1)

ggsave(
  plot = multiple_page_layout,
  filename = paste0(out_path,time_stamp(),"plot_UMAP_Cycle.pdf"),
  device = "pdf",
  width = 210,
  height = 297,
  unit = "mm")

```

# Export files

```{r Clean the objects}

# Keep only useful columns in metadata
columns_to_keep = c("Sample_Description", "cell_cycle_phase", "detected","subsets_mitochondrial_reads_percent","scDblFinder.score")
columns_to_keep = which(colnames(colData(sce)) %in% columns_to_keep)
colData(sce) = colData(sce)[columns_to_keep]

```

# Exporet files

```{r export the objects}

# Export annotated sce object
saveRDS(
  object = sce,
  file = paste0(out_path,time_stamp(),"sce_cycle_annotated.rds")
  )


```
